Refine item thumbnail helpers
